Add conditoinal woodlotOperationsTxt and refactor appUtils.ts
