+++ admin/src/app/foms/summary/summary.component.html
@@ -168,7 +168,12 @@
         </section>
 
         <section *ngIf="!selectedScope?.commentScopeCode">
-          <small>This FOM can be relied upon by the FOM holder for the purpose of a cutting permit or road permit application, until the date three years after commencement of the public review and commenting period. FOMs published by BC Timber Sales can be relied upon for the purpose of a cutting permit or road permit application, or the issuance of a Timber Sales License until the date three years after conclusion of the public review and commenting period.</small>
+          <small>
+            {{project.projectPlanCode == projectPlanCodeEnum.Fsp
+                ? periodOperationsTxt
+                : woodlotOperationsTxt
+            }}
+          </small>
         </section>
 
       </section>
