+++ admin/src/app/foms/summary/summary.component.ts
@@ -9,10 +9,10 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
 import { ActivatedRoute, RouterLink } from '@angular/router';
 import {
-  AttachmentResponse, AttachmentService, InteractionResponse, InteractionService,
-  ProjectPlanCodeEnum,
-  ProjectResponse, ProjectService, PublicCommentAdminResponse, PublicCommentService,
-  SpatialFeaturePublicResponse, SpatialFeatureService
+    AttachmentResponse, AttachmentService, InteractionResponse, InteractionService,
+    ProjectPlanCodeEnum,
+    ProjectResponse, ProjectService, PublicCommentAdminResponse, PublicCommentService,
+    SpatialFeaturePublicResponse, SpatialFeatureService
 } from '@api-client';
 import { ConfigService } from '@utility/services/config.service';
 import * as _ from 'lodash';
@@ -46,6 +46,8 @@ import { InteractionsSummaryComponent } from './interactions-summary/interaction
 })
 export class SummaryComponent implements OnInit, OnDestroy {
   readonly projectPlanCodeEnum = ProjectPlanCodeEnum;
+  readonly periodOperationsTxt = "This FOM can be relied upon by the FOM holder for the purpose of a cutting permit or road permit application, until the date three years after commencement of the public review and commenting period. FOMs published by BC Timber Sales can be relied upon for the purpose of a cutting permit or road permit application, or the issuance of a Timber Sales License until the date three years after conclusion of the public review and commenting period.";
+  readonly woodlotOperationsTxt = "Woodlots are not legally required to publish FOMs for public review and comment prior to cutting permit or road permit application. However, woodlot licensees may choose to publish FOMs on a voluntary basis to facilitate public engagement.";
   projectId: number;
   project: ProjectResponse;
   projectReqError: boolean;
