+++ public/src/app/applications/app-public-notices/public-notices-panel.component.html
@@ -192,7 +192,10 @@
       </div>
 
       <div class="validity-txt" *ngIf="isFomAvailable">
-        This FOM can be relied upon by the FOM holder for the purpose of a cutting permit or road permit application, until the date three years after commencement of the public review and commenting period. FOMs published by BC Timber Sales can be relied upon for the purpose of a cutting permit or road permit application, or the issuance of a Timber Sales License until the date three years after conclusion of the public review and commenting period.
+        {{pn.project.projectPlanCode == projectPlanCodeEnum.Fsp
+            ? periodOperationsTxt
+            : woodlotOperationsTxt
+        }}
       </div>
 
     </ng-template>
