+++ public/src/app/applications/details-panel/details-panel.component.ts
@@ -2,14 +2,15 @@ import { CommonModule } from '@angular/common';
 import { Component, ElementRef, EventEmitter, OnDestroy, OnInit, Output, ViewChild } from '@angular/core';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import {
-  AttachmentResponse, AttachmentService, ProjectPlanCodeEnum, ProjectResponse, ProjectService,
-  SpatialFeaturePublicResponse, SpatialFeatureService, WorkflowStateCode
+    AttachmentResponse, AttachmentService, ProjectPlanCodeEnum, ProjectResponse, ProjectService,
+    SpatialFeaturePublicResponse, SpatialFeatureService, WorkflowStateCode
 } from '@api-client';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { faArrowUpRightFromSquare } from '@fortawesome/free-solid-svg-icons';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
+import { periodOperationsTxt, woodlotOperationsTxt } from '@public-core/constants/appConstants';
 import { UrlService } from '@public-core/services/url.service';
-import { getCommentingClosingDate } from '@public-core/utils/constants/appUtils';
+import { getCommentingClosingDate } from '@public-core/utils/appUtils';
 import { ConfigService } from '@utility/services/config.service';
 import { FeatureSelectService } from '@utility/services/featureSelect.service';
 import { DetailsMapComponent } from 'app/applications/details-panel/details-map/details-map.component';
@@ -59,10 +60,10 @@ export class DetailsPanelComponent implements OnDestroy, OnInit {
   public attachments: AttachmentResponse[];
   public faArrowUpRightFromSquare = faArrowUpRightFromSquare;
   public getCommentingClosingDate = getCommentingClosingDate;
-  
-  public periodOperationsTxt = "This FOM can be relied upon by the FOM holder for the purpose of a cutting permit or road permit application, until the date three years after commencement of the public review and commenting period. FOMs published by BC Timber Sales can be relied upon for the purpose of a cutting permit or road permit application, or the issuance of a Timber Sales License until the date three years after conclusion of the public review and commenting period."
-  public periodOperationsTooltipTxt = "An FSP holder has three years to apply for a cutting permit or road permit for cutblocks and roads displayed on a FOM. This is called the validity period, it starts on the day commenting opens on a FOM. For BC Timber Sales the validity period starts on the day commenting closes."
+  public periodOperationsTooltipTxt = "An FSP holder has three years to apply for a cutting permit or road permit for cutblocks and roads displayed on a FOM. This is called the validity period, it starts on the day commenting opens on a FOM. For BC Timber Sales the validity period starts on the day commenting closes.";
   readonly projectPlanCodeEnum = ProjectPlanCodeEnum;
+  readonly periodOperationsTxt = periodOperationsTxt;
+  readonly woodlotOperationsTxt = woodlotOperationsTxt;
 
   constructor(
     public modalService: NgbModal,
