+++ public/src/core/services/fomFilters.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { AppUtils } from '@public-core/utils/constants/appUtils';
+import { AppUtils } from '@public-core/utils/appUtils';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { Filter, IFilter, IMultiFilter, IMultiFilterFields, MultiFilter } from '../../app/applications/utils/filter';
