+++ public/src/core/utils/appUtils.ts
